reorganize color module
