+++ .gitconfig
@@ -1,0 +1,19 @@
+[diff]
+
+    tool = meld
+
+[difftool]
+
+    prompt = false
+
+[difftool "meld"]
+
+    cmd = meld "$LOCAL" "$REMOTE"
+
+[merge]
+
+    tool = meld
+
+[mergetool "meld"]
+
+    cmd = meld "$LOCAL" "$MERGED" "$REMOTE" --output "$MERGED"
