+++ torchgeometry/color/__init__.py
@@ -1,18 +1,22 @@
-from ._color import rgb_to_grayscale, RgbToGrayscale
-from ._color import BgrToRgb, bgr_to_rgb
-from ._color import RgbToBgr, rgb_to_bgr
-from ._color import RgbToHsv, rgb_to_hsv
-from ._color import Normalize, normalize
+from .gray import rgb_to_grayscale, RgbToGrayscale
+from .rgb import BgrToRgb, bgr_to_rgb
+from .rgb import RgbToBgr, rgb_to_bgr
+from .hsv import RgbToHsv, rgb_to_hsv
+from .hsv import HsvToRgb, hsv_to_rgb
+from .normalize import Normalize, normalize
 
 __all__ = [
     "rgb_to_grayscale",
+    "bgr_to_rgb",
+    "rgb_to_bgr",
+    "rgb_to_hsv",
+    "hsv_to_rgb",
+    "normalize"
     "RgbToGrayscale",
     "BgrToRgb",
-    "bgr_to_rgb",
     "RgbToBgr",
-    "rgb_to_bgr",
     "RgbToHsv",
-    "rgb_to_hsv",
+    "HsvToRgb",
     "Normalize",
-    "normalize"
+
 ]
